Add preventBackAndClose
