--- packages/hello-gsm/src/hooks/usePreventClose.ts
@@ -1,15 +1,0 @@
-import { useEffect } from 'react';
-
-const preventClose = (e: BeforeUnloadEvent) => {
-  e.preventDefault();
-  e.returnValue = '';
-};
-
-export const usePreventClose = () =>
-  useEffect(() => {
-    window.addEventListener('beforeunload', preventClose);
-
-    return () => {
-      window.removeEventListener('beforeunload', preventClose);
-    };
-  }, []);
